make a new component for input box
