+++ package.json
@@ -17,7 +17,7 @@
     "@ai-sdk/google": "^1.2.22",
     "@ai-sdk/openai": "^1.3.22",
     "@ai-sdk/vue": "^1.2.12",
-    "@beekeeperstudio/plugin": "^1.4.0-beta.5",
+    "@beekeeperstudio/plugin": "^1.4.0-beta.6",
     "@langchain/core": "^0.3.61",
     "@material-symbols/font-400": "^0.31.2",
     "ai": "^4.3.16",
