+++ src/components/ChatInterface.vue
@@ -60,46 +60,8 @@
       <div ref="bottomMarker"></div>
     </div>
     <div class="chat-input-container-container">
-      <div class="chat-input-container">
-        <BaseInput
-          type="textarea"
-          v-model="input"
-          @keydown.enter="handleEnterKey"
-          @keydown.up="handleUpArrow"
-          @keydown.down="handleDownArrow"
-          placeholder="Type your message here"
-          rows="1"
-        />
-        <div class="actions">
-          <Dropdown
-            :model-value="model"
-            placeholder="Select Model"
-            aria-label="Model"
-          >
-            <DropdownOption
-              v-for="optionModel in filteredModels"
-              :key="optionModel.id"
-              :value="optionModel.id"
-              :text="optionModel.id"
-              :selected="matchModel(optionModel, model)"
-              @select="selectModel(optionModel)"
-            />
-            <div class="dropdown-separator"></div>
-            <button class="dropdown-action" @click="$emit('manage-models')">
-              Manage models
-            </button>
-          </Dropdown>
-          <button
-            v-if="canSendMessage"
-            @click="submit"
-            class="submit-btn"
-            :disabled="!input.trim()"
-          >
-            <span class="material-symbols-outlined">send</span>
-          </button>
-          <button v-else @click="stop" class="stop-btn" />
-        </div>
-      </div>
+      <PromptInput storage-key="inputHistory" :processing="processing" :selected-model="model"
+       @select-model="selectModel" @manage-models="$emit('manage-models')" @submit="submit" @stop="stop"  />
     </div>
   </div>
 </template>
@@ -118,10 +80,8 @@ import { PropType } from "vue";
 import { mapActions, mapGetters, mapState, mapWritableState } from "pinia";
 import { RootBinding } from "@/plugins/appEvent";
 import { useInternalDataStore } from "@/stores/internalData";
-import { matchModel } from "@/utils";
 import BaseInput from "@/components/common/BaseInput.vue";
-
-const maxHistorySize = 50;
+import PromptInput from "@/components/common/PromptInput.vue";
 
 export default {
   name: "ChatInterface",
@@ -133,6 +93,7 @@ export default {
     ToolMessage,
     Markdown,
     BaseInput,
+    PromptInput,
   },
 
   emits: ["manage-models", "open-configuration"],
@@ -170,12 +131,7 @@ export default {
   },
 
   data() {
-    const inputHistoryStr = localStorage.getItem("inputHistory") || "[]";
-    const inputHistory: string[] = JSON.parse(inputHistoryStr);
-    inputHistory.push("");
     return {
-      inputHistory,
-      inputIndex: inputHistory.length - 1,
       isAtBottom: true,
       showFullError: false,
       noModelError: false,
@@ -185,22 +141,9 @@ export default {
   computed: {
     ...mapGetters(useChatStore, ["systemPrompt"]),
     ...mapWritableState(useChatStore, ["model"]),
-    ...mapState(useChatStore, {
-      filteredModels(store) {
-        return store.models.filter((m) => m.enabled);
-      },
-    }),
-    input: {
-      get() {
-        return this.inputHistory[this.inputIndex];
-      },
-      set(value: string) {
-        this.inputHistory[this.inputIndex] = value;
-      },
-    },
-    canSendMessage() {
-      if (this.askingPermission && this.input.trim().length > 0) return true;
-      return this.status === "ready" || this.status === "error";
+    processing() {
+      if (this.askingPermission) return false;
+      return this.status !== "ready" && this.status !== "error";
     },
     showSpinner() {
       return (
@@ -270,137 +213,25 @@ export default {
 
   methods: {
     ...mapActions(useInternalDataStore, ["setInternal"]),
-    matchModel,
-    handleEnterKey(e) {
-      if (e.shiftKey) {
-        // Allow default behavior (new line) when Shift+Enter is pressed
-        return;
-      }
 
-      if (this.canSendMessage) {
-        e.preventDefault();
-        e.stopPropagation();
-        this.submit();
-      }
-    },
-
-    // Handle up/down arrow keys for history navigation
-    handleUpArrow(e) {
-      const textarea = e.target;
-      const text = textarea.value;
-
-      // Is cursor at first line?
-      const cursorPos = textarea.selectionStart;
-      const textBeforeCursor = text.substring(0, cursorPos);
-
-      // If there's no newline before cursor or cursor is at position 0, we're at the first line
-      if (cursorPos === 0 || textBeforeCursor.lastIndexOf("\n") === -1) {
-        // Go back in history
-        if (this.navigateHistory(-1)) {
-          e.preventDefault();
-        }
-      }
-    },
-
-    handleDownArrow(e) {
-      const textarea = e.target;
-      const text = textarea.value;
-
-      // Is cursor at last line?
-      const cursorPos = textarea.selectionStart;
-      const textAfterCursor = text.substring(cursorPos);
-
-      // If there's no newline after cursor or cursor is at end of text, we're at the last line
-      if (cursorPos === text.length || textAfterCursor.indexOf("\n") === -1) {
-        // Go forward in history
-        if (this.navigateHistory(1)) {
-          e.preventDefault();
-        }
-      }
-    },
-
-    /** Navigate through input history. Returns true if input changed. */
-    navigateHistory(direction: 1 | -1) {
-      const oldIndex = this.inputIndex;
-
-      this.inputIndex = _.clamp(
-        this.inputIndex + direction,
-        0,
-        this.inputHistory.length - 1
-      );
-
-      const changed = this.inputIndex !== oldIndex;
-
-      if (changed) {
-        // Place cursor at the end of the input text
-        this.$nextTick(() => {
-          const textarea = document.querySelector("textarea");
-          if (textarea) {
-            textarea.selectionStart = textarea.selectionEnd =
-              textarea.value.length;
-          }
-        });
-      }
-
-      return changed;
-    },
-
-    submit() {
-      // Don't add to history if the message starts with a space
-      const preventHistory = this.input.startsWith(" ");
-      const message = this.input.trim();
-
-      // Don't send empty messages
-      if (!message) return;
-
+    submit(input: string) {
       if (!this.model) {
         // FIXME we should catch this and show it on screen
         this.noModelError = true;
         return;
       }
 
-      if (!preventHistory) {
-        this.addToHistory(message);
-      }
-
       this.noModelError = false;
-      this.resetInput();
 
       if (this.askingPermission) {
-        this.rejectPermission(message);
+        this.rejectPermission(input);
       } else {
-        this.send(message, this.getSendOptions());
+        this.send(input, this.getSendOptions());
       }
     },
 
     async reload() {
       await this.retry(this.getSendOptions());
-    },
-
-    resetInput() {
-      if (this.inputHistory[this.inputHistory.length - 1] === "") {
-        this.inputIndex = this.inputHistory.length - 1;
-      } else {
-        this.inputHistory.push("");
-        this.inputIndex = this.inputHistory.length - 1;
-      }
-    },
-
-    addToHistory(input: string) {
-      const oldHistory = JSON.parse(localStorage.getItem("inputHistory")!);
-
-      let newHistory = [...oldHistory, input];
-
-      // Limit history size
-      if (newHistory.length > maxHistorySize) {
-        newHistory = newHistory.slice(-maxHistorySize);
-      }
-
-      localStorage.setItem("inputHistory", JSON.stringify(newHistory));
-
-      // Reset history navigation
-      this.inputHistory = newHistory;
-      this.inputIndex = newHistory.length - 1;
     },
 
     stop() {
@@ -418,6 +249,7 @@ export default {
       this.$refs.scrollContainerRef.scrollTop =
         this.$refs.scrollContainerRef.scrollHeight;
     },
+
     selectModel(model: Model) {
       this.setInternal("lastUsedModelId", model.id);
       this.model = model;
